lab-1-instrumentation/pyproject.toml
[tool.poetry]
name = "lab-1-instrumentation"
version = "0.1.0"
description = ""
authors = ["Erasmo <erasmolpa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
opentelemetry-distro = "*"
opentelemetry-exporter-otlp = "*"
opentelemetry-api = "*"
opentelemetry-sdk = "*"
opentelemetry-instrumentation-flask = "*"
opentelemetry-exporter-jaeger = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app:main"